Added request simulation to request_test.py
